final breakdown of the extension.ts file
