+++ packages/zowe-explorer/src/SpoolProvider.ts
@@ -67,12 +67,12 @@ export function encodeJobFile(session: string, spool: zowe.IJobFile): vscode.Uri
  */
 export const toUniqueJobFileUri =
     (session: string, spool: zowe.IJobFile) =>
-    (uniqueFragment: string): vscode.Uri => {
-        const encodedUri = encodeJobFile(session, spool);
-        return encodedUri.with({
-            fragment: uniqueFragment,
-        });
-    };
+        (uniqueFragment: string): vscode.Uri => {
+            const encodedUri = encodeJobFile(session, spool);
+            return encodedUri.with({
+                fragment: uniqueFragment,
+            });
+        };
 
 /**
  * Decode the information needed to get the Spool content.
@@ -82,4 +82,12 @@ export const toUniqueJobFileUri =
 export function decodeJobFile(uri: vscode.Uri): [string, zowe.IJobFile] {
     const [session, spool] = JSON.parse(uri.query) as [string, zowe.IJobFile];
     return [session, spool];
+}
+
+export function initializeSpoolProvider(context: vscode.ExtensionContext) {
+    const spoolProvider = new SpoolProvider();
+    const providerRegistration = vscode.Disposable.from(
+        vscode.workspace.registerTextDocumentContentProvider(SpoolProvider.scheme, spoolProvider)
+    );
+    context.subscriptions.push(spoolProvider, providerRegistration);
 }
