+++ packages/zowe-explorer/src/utils/LoggerUtils.ts
@@ -1,0 +1,35 @@
+/*
+ * This program and the accompanying materials are made available under the terms of the *
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at *
+ * https://www.eclipse.org/legal/epl-v20.html                                      *
+ *                                                                                 *
+ * SPDX-License-Identifier: EPL-2.0                                                *
+ *                                                                                 *
+ * Copyright Contributors to the Zowe Project.                                     *
+ *                                                                                 *
+ */
+
+import * as vscode from "vscode";
+import * as nls from "vscode-nls";
+import * as globals from "../globals";
+
+// Set up localization
+nls.config({
+    messageFormat: nls.MessageFormat.bundle,
+    bundleFormat: nls.BundleFormat.standalone,
+})();
+const localize: nls.LocalizeFunc = nls.loadMessageBundle();
+
+export async function initializeZoweLogger(context: vscode.ExtensionContext): Promise<void> {
+    try {
+        globals.initLogger(context);
+        globals.LOG.debug(localize("initialize.log.debug", "Initialized logger from VSCode extension"));
+    } catch (err) {
+        globals.LOG.error(err);
+        const errorMessage = localize(
+            "initialize.log.error",
+            "Error encountered while activating and initializing logger! "
+        );
+        await vscode.window.showErrorMessage(`${errorMessage}: ${err.message}`);
+    }
+}
